requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
coverage==5.5
idna==3.1
iniconfig==1.1.1
multidict==5.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
pytz==2020.5
style==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
update==0.0.1
yarl==1.6.3
